+++ DQM/SiTrackerPhase2/plugins/Phase2TrackerMonitorDigi.cc
@@ -36,12 +36,12 @@
 #include "DataFormats/SiPixelDigi/interface/PixelDigi.h"
 #include "DataFormats/SiPixelDigi/interface/PixelDigiCollection.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "SimTracker/SiPhase2Digitizer/plugins/Phase2TrackerDigitizerFwd.h"
 
 // DQM Histograming
 #include "DQMServices/Core/interface/MonitorElement.h"
 #include "DQM/SiTrackerPhase2/interface/TrackerPhase2DQMUtil.h"
 
+using Phase2TrackerGeomDetUnit = PixelGeomDetUnit;
 //
 // constructors
 //
