+++ GeneratorInterface/BeamHaloGenerator/interface/chepevt.inc
@@ -1,5 +1,5 @@
 C -- see in /afs/cern.ch/cms/sw/slc3_ia32_gcc323/external/clhep/1.9.2.2/include/CLHEP/
 
 
-#include "SimDataFormats/GeneratorProducts/data/stdhep.inc"
-#include "SimDataFormats/GeneratorProducts/data/hepev4.inc"
+#include "SimDataFormats/GeneratorProducts/interface/stdhep.inc"
+#include "SimDataFormats/GeneratorProducts/interface/hepev4.inc"
