+++ GeneratorInterface/Hydjet2Interface/src/HepEvt.F
@@ -1,8 +1,8 @@
 
       subroutine ihepevt
   
-#include "SimDataFormats/GeneratorProducts/data/stdhep.inc"
-#include "SimDataFormats/GeneratorProducts/data/hepev4.inc"
+#include "SimDataFormats/GeneratorProducts/interface/stdhep.inc"
+#include "SimDataFormats/GeneratorProducts/interface/hepev4.inc"
   
       return
       end
