+++ SimMuon/DTDigitizer/test/DTDigiAnalyzer.cc
@@ -31,7 +31,7 @@
 
 #include "TFile.h"
 
-#include "SimMuon/DTDigitizer/test/Histograms.h"
+#include "SimMuon/DTDigitizer/interface/Histograms.h"
 
 using namespace edm;
 using namespace std;
