+++ Validation/DTRecHits/test/writeSummaryTable.r
@@ -27,7 +27,7 @@ void writeSummaryTable(TString filename, bool doEff=true, bool doEffab = true, b
   //----------------------------------------------------------------------
 
   if (! TString(gSystem->GetLibraries()).Contains("Histograms_h")) {
-    gROOT->LoadMacro("$CMSSW_BASE/src/Validation/DTRecHits/test/Histograms.h+");
+    gROOT->LoadMacro("$CMSSW_BASE/src/Validation/DTRecHits/interface/Histograms.h+");
     gROOT->LoadMacro("macros.C");
     gROOT->LoadMacro("ranges.C");
   }
